+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Loading thumbnails into textures
 - A proxy video setting, and the ability to try with proxying, if the video errors on the first try.
+- A PlayFromScript sample
 
 ## [3.2.0] - 2024-04-13